Add --cat/-c option to process-incoming
When passed, the message text is output to STDOUT again, and the
(interactive) status message is output to STDERR instead.

This allows to use muacrypt in a (Neo)Mutt display-filter. For example,

	set display_filter='muacrypt process-incoming --cat'

as suggested in [0].

[0] https://www.mail-archive.com/mutt-users@mutt.org/msg22068.html

Signed-off-by: Olivier Mehani <shtrom@ssji.net>

squash! Add --cat/-c option to process-incoming

Capture ValueError on missing Delivered-To, and display a nice error
message instead (related to #69).

Signed-off-by: Olivier Mehani <shtrom@ssji.net>